:recycle: article周りのコンポーネント化
