+++ src/app/article/[slug]/page.jsx
@@ -1,7 +1,9 @@
-import { fetchArticle } from "@/app/lib/data";
-import DeleteArticle from "@/app/ui/article/delete-article";
 import { cookies } from "next/headers";
-import Link from "next/link";
+import { fetchArticle } from "@/app/lib/data";
+import ArticleMeta from "@/app/ui/articles/article-meta";
+import EditArticle from "@/app/ui/articles/article/edit-article";
+import DeleteArticle from "@/app/ui/articles/article/delete-article";
+import ArticleContent from "@/app/ui/articles/article/article-content";
 
 export default async function Page({ params }) {
   const slug = params.slug;
@@ -16,23 +18,10 @@ export default async function Page({ params }) {
             <h1>{article.title}</h1>
 
             <div className="article-meta">
-              <Link href="/profile/eric-simons">
-                <img src="http://i.imgur.com/Qr71crq.jpg" />
-              </Link>
-              <div className="info">
-                <Link href="/profile/eric-simons" className="author">
-                  {article.author.username}
-                </Link>
-                <span className="date">January 20th</span>
-              </div>
+              <ArticleMeta article={article} />
               {user && user.value == article.author.username && (
                 <>
-                  <Link
-                    href={`/editor/${slug}`}
-                    class="btn btn-sm btn-outline-secondary"
-                  >
-                    <i class="ion-edit"></i> Edit Article
-                  </Link>
+                  <EditArticle slug={slug} />
                   <DeleteArticle slug={slug} />
                 </>
               )}
@@ -40,20 +29,7 @@ export default async function Page({ params }) {
           </div>
         </div>
 
-        <div className="container page">
-          <div className="row article-content">
-            <div className="col-md-12">
-              <p>{article.description}</p>
-              <p>{article.body}</p>
-              <ul className="tag-list">
-                {article.tagList.map((tag) => (
-                  <li className="tag-default tag-pill tag-outline">{tag}</li>
-                ))}
-              </ul>
-            </div>
-          </div>
-          <hr />
-        </div>
+        <ArticleContent article={article} />
       </div>
     </>
   );
