+++ src/app/layout.js
@@ -1,13 +1,8 @@
 import "@/app/global.css";
+import { authenticated } from "@/app/lib/auth";
 import Header from "@/app/ui/header";
 import HeaderWithLogin from "@/app/ui/header-with-login";
 import Footer from "@/app/ui/footer";
-import { authenticated } from "@/app/lib/auth";
-
-// export const metadata = {
-//   title: "Create Next App",
-//   description: "Generated by create next app",
-// };
 
 export default async function RootLayout({ children }) {
   return (
