+++ src/app/ui/articles/article/article-content.jsx
@@ -1,0 +1,20 @@
+export default function ArticleContent({ article }) {
+  return (
+    <>
+      <div className="container page">
+        <div className="row article-content">
+          <div className="col-md-12">
+            <p>{article.description}</p>
+            <p>{article.body}</p>
+            <ul className="tag-list">
+              {article.tagList.map((tag) => (
+                <li className="tag-default tag-pill tag-outline">{tag}</li>
+              ))}
+            </ul>
+          </div>
+        </div>
+        <hr />
+      </div>
+    </>
+  );
+}
