+++ src/app/ui/articles/feed/feed-content.jsx
@@ -1,0 +1,18 @@
+import Link from "next/link";
+
+export default function FeedContent({ article }) {
+  return (
+    <>
+      <Link href={`/article/${article.slug}`} className="preview-link">
+        <h1>{article.title}</h1>
+        <p>{article.description}</p>
+        <span>Read more...</span>
+        <ul className="tag-list">
+          {article.tagList.map((tag) => (
+            <li className="tag-default tag-pill tag-outline">{tag}</li>
+          ))}
+        </ul>
+      </Link>
+    </>
+  );
+}
