+++ src/app/ui/articles/feed/feed-toggle.jsx
@@ -1,0 +1,13 @@
+import Link from "next/link";
+
+export default function FeedToggle({ feedName, active = "" }) {
+  return (
+    <>
+      <li className="nav-item">
+        <Link className={`nav-link ${active}`} href="">
+          {feedName}
+        </Link>
+      </li>
+    </>
+  );
+}
