+++ src/app/ui/top/banner.jsx
@@ -1,0 +1,12 @@
+export default function Banner() {
+  return (
+    <>
+      <div className="banner">
+        <div className="container">
+          <h1 className="logo-font">conduit</h1>
+          <p>A place to share your knowledge.</p>
+        </div>
+      </div>
+    </>
+  );
+}
